+++ src/api/ICaselabEcmApi.ts
@@ -1,9 +1,9 @@
-import type IOrganisation from './IOrganisation';
-import type IUser from './IUser';
+import type IOrganisation from '../Entities/IOrganization';
+import type IUser from '../Entities/IUser';
 
 export default interface ICaselabEcmApi {
   loginService: (email: string, password: string) => Promise<string>;
   getUsersCount: (token: string) => Promise<string>;
   getAllUsers: (token: string) => Promise<IUser[]>;
-  getAllOrganization: (token: string) => Promise<IOrganisation[]>;
+  getAllOrganizations: (token: string) => Promise<IOrganisation[]>;
 }
